boggle-backend/requirements.txt
# Core dependencies
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.25

# Testing
pytest==8.0.2

# CORS handling (for frontend/backend communication)
Flask-Cors==4.0.0

# Optional but recommended for development
python-dotenv==1.0.0  # For environment variables